// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        ExtractFlowDocumentResponse: {
            document: components["schemas"]["FlowDocument"];
            /** Errors */
            errors: components["schemas"]["ModuleErrorDTO"][];
        };
// ...
        FlowDocument: {
            /** Text */
            text: string;
            /** Blocks */
            blocks: components["schemas"]["FlowBlock"][];
        };
// ...
        PageDocument: {
            /** Pages */
            pages: components["schemas"]["DocumentPage"][];
        };
// ...
        ResumeValidationPolicy: {
            maximumFileSizeBytes: components["schemas"]["FileSizeLimits"];
            /** Maximumdirecttextcharacters */
            maximumDirectTextCharacters: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}